Add CI badge
Add a CI badge to let the users know about the health of the project.
Unfortunately, Github Actions cannot create a badge for two different
events, e.g., pushes to `master` and nightly runs. Of the two events,
the second is probably more helpful to users, so we opt to show that.